Sync Updates

+++ hoomd/CommunicatorGPU.cuh
@@ -58,7 +58,7 @@ void gpu_stage_particles(const unsigned int n,
     \param d_neighbors List of neighbor ranks
     \param alloc Caching allocator
  */
-void gpu_sort_migrating_particles(const unsigned int nsend,
+void gpu_sort_migrating_particles(const size_t nsend,
                    pdata_element *d_in,
                    const unsigned int *d_comm_flags,
                    const Index3D& di,
